This PR is based on top of https://github.com/rust-lang-nursery/rand/pull/233.

Parts of the changes here come from my attempts with a with `BlockRngCore` trait, like https://github.com/rust-lang-nursery/rand/pull/242. And better to finish this seperately, so all functional changes and changes to the documentation can be reviewed better.

The `init` function contained useful documentation that was not visible because `init` was not a public function. So I moved that to the documentation of ChaChaRng. Also I wrote some more in a similar style as HC-128.

I added some tests and better documentation for `set_counter`. Trying to use it to set a nonce is difficult, considering the trouble with endianness and different standards specifying different schemes for the nonce and counter. So I documented it to be not a supported use of `set_rounds`.

It turned out to be easy to make the number of rounds the core ChaCha algorithm runs configurable, just the addition of a `set_rounds` method.

And I stumbled on a way to improve the performance, by using an `usize` for the number of rounds and keeping the core algorithm a seperate function. No idea why it works though.

Benchmarks before:
```
test gen_bytes_chacha   ... bench:   2,782,004 ns/iter (+/- 5,165) = 368 MB/s
test gen_u32_chacha     ... bench:      10,774 ns/iter (+/- 11) = 371 MB/s
test gen_u64_chacha     ... bench:      21,564 ns/iter (+/- 75) = 370 MB/s
```

Benchmarks after this and https://github.com/rust-lang-nursery/rand/pull/242:
```
test gen_bytes_chacha   ... bench:   1,898,515 ns/iter (+/- 6,578) = 539 MB/s
test gen_u32_chacha     ... bench:       8,357 ns/iter (+/- 24) = 478 MB/s
test gen_u64_chacha     ... bench:      16,207 ns/iter (+/- 47) = 493 MB/s
```

Finally remembered why it is a good idea to wait with generating results until the first use: it makes things like `ChaChaRng::new().set_counter(500)` possible.

cc @burdges, because I know you have an interest in ChaCha.