@pitdicker 2018-01-21
Good comments, thank you.

> how does ChaCha with 8 or 12 rounds compare to HC128 and ISAAC in terms of performance?

Still slowest, but not bad at all:
```
test gen_bytes_chacha   ... bench:     902,419 ns/iter (+/- 15,331) = 1134 MB/s
test gen_bytes_hc128    ... bench:     441,192 ns/iter (+/- 2,454) = 2320 MB/s
test gen_bytes_isaac    ... bench:     710,000 ns/iter (+/- 2,664) = 1442 MB/s
test gen_bytes_isaac64  ... bench:     381,510 ns/iter (+/- 870) = 2684 MB/s
```