@pitdicker 2018-01-24
I tried using the code from @PeterReid's implementation of ChaCha, integrating it was super easy.

Results on stable:
```
test gen_bytes_chacha20 ... bench:   2,088,500 ns/iter (+/- 3,958) = 490 MB/s
test gen_u32_chacha20   ... bench:       9,195 ns/iter (+/- 31) = 435 MB/s
test gen_u64_chacha20   ... bench:      18,828 ns/iter (+/- 67) = 424 MB/s
```

Results on nightly with `#[repr(simd)]`:
```
test gen_bytes_chacha20 ... bench:   1,796,552 ns/iter (+/- 4,974) = 569 MB/s
test gen_u32_chacha20   ... bench:       7,722 ns/iter (+/- 24) = 518 MB/s
test gen_u64_chacha20   ... bench:      15,510 ns/iter (+/- 90) = 515 MB/s
```

Compared with our current implementation:
```
test gen_bytes_chacha   ... bench:   1,898,515 ns/iter (+/- 6,578) = 539 MB/s
test gen_u32_chacha     ... bench:       8,357 ns/iter (+/- 24) = 478 MB/s
test gen_u64_chacha     ... bench:      16,207 ns/iter (+/- 47) = 493 MB/s
```

The result with SIMD can be improved again with ±3% by making the state also `#[repr(simd)]`. The assembly for SIMD looks pretty much optimal to me. Yet it regresses stable by another 12%. 

I am not sure if this change is worth it at the moment, because it is only faster on nightly. But nice to know that our implementation is actually pretty decent already.