@dhardy 2018-01-25
Think I'll go ahead and merge manually. It's clear from the Travis failures that GitHub messed up the merge :-(